Relax the dependency on `readline`
`readline` is not available on Windows. All it provides is tab-completion of base commands, which is dubiously useful. So, let's not crash the entire program if it's missing, that would be silly.

If `pyreadline3` is available, that's an OK fallback on Windows, so let's try it quickly before moving on.

This should clear the only major blocker to Windows support.