Catch critically outdated `drozer-agent`
Some people install `drozer-agent` from dodgy sources, and they often end up installing v2.x.x as a result. This really breaks things, because we changed the bundle identifier in drozer 3.

Since it's a common enough issue (for now), let's catch the exception this generates and direct users to the `drozer-agent` repo.