`ModuleLoader` bugfixes
When the `DROIDHG_MODULE_PATH` was retired, the list of repositories carried on being presented as a `:`-delimited string, for seemingly no good reason. That string was built, and then immediately split to get a list.

This also introduced bugs in Windows, since every absolute path will contain a `:`. Consequently, the loader only worked sometimes (by accident) on Windows, and it usually tried to load a few irrelevant directories as repos.

Let's not do any of that and just pass lists around.