Fix `app.provider.info`
If the authority for a content provider came back as null, it would be represented as a `ReflectedNull` in drozer. `ReflectedNull` doesn't have a `split()` function like strings do, so this caused a crash.

Apparently this was OK in Python 2 because it would get automatically typecast as a `str`. But let's do things properly here and actually handle the null case manually.